moving cli to integration/cli
