+++ tests/integration/cli/test_describe_downloads.py
@@ -7,8 +7,7 @@ from unittest.case import skipIf
 from datumaro.components.extractor_tfds import AVAILABLE_TFDS_DATASETS, TFDS_EXTRACTOR_AVAILABLE
 from datumaro.util import parse_json
 
-from ..requirements import Requirements, mark_requirement
-
+from tests.requirements import Requirements, mark_requirement
 from tests.utils.test_utils import TestDir, mock_tfds_data
 from tests.utils.test_utils import run_datum as run
