+++ tests/integration/cli/test_detect_format.py
