+++ tests/integration/cli/test_filter.py
@@ -8,16 +8,15 @@ from datumaro.components.extractor import DatasetItem
 from datumaro.components.project import Project
 from datumaro.util.scope import scope_add, scoped
 
-from ..requirements import Requirements, mark_requirement
-
+from tests.requirements import Requirements, mark_requirement
 from tests.utils.test_utils import TestDir, compare_datasets
 from tests.utils.test_utils import run_datum as run
 
 
-class TransformTest(TestCase):
+class FilterTest(TestCase):
     @mark_requirement(Requirements.DATUM_GENERAL_REQ)
     @scoped
-    def test_can_transform_dataset_inplace(self):
+    def test_can_filter_dataset_inplace(self):
         test_dir = scope_add(TestDir())
         Dataset.from_iterable(
             [
@@ -27,33 +26,20 @@ class TransformTest(TestCase):
             categories=["a", "b"],
         ).export(test_dir, "coco")
 
-        run(
-            self,
-            "transform",
-            "-t",
-            "remap_labels",
-            "--overwrite",
-            test_dir + ":coco",
-            "--",
-            "-l",
-            "a:cat",
-            "-l",
-            "b:dog",
-        )
+        run(self, "filter", "-e", '/item[id = "1"]', "--overwrite", test_dir + ":coco")
 
         expected_dataset = Dataset.from_iterable(
             [
                 DatasetItem(1, annotations=[Label(0, id=1, group=1)]),
-                DatasetItem(2, annotations=[Label(1, id=2, group=2)]),
             ],
-            categories=["cat", "dog"],
+            categories=["a", "b"],
         )
         compare_datasets(
             self, expected_dataset, Dataset.import_from(test_dir, "coco"), ignored_attrs="*"
         )
 
     @mark_requirement(Requirements.DATUM_GENERAL_REQ)
-    def test_transform_fails_on_inplace_update_without_overwrite(self):
+    def test_filter_fails_on_inplace_update_without_overwrite(self):
         with TestDir() as test_dir:
             Dataset.from_iterable(
                 [
@@ -62,10 +48,10 @@ class TransformTest(TestCase):
                 categories=["a", "b"],
             ).export(test_dir, "coco")
 
-            run(self, "transform", "-t", "random_split", test_dir + ":coco", expected_code=1)
+            run(self, "filter", "-e", "/item", test_dir + ":coco", expected_code=1)
 
     @mark_requirement(Requirements.DATUM_GENERAL_REQ)
-    def test_transform_fails_on_inplace_update_of_stage(self):
+    def test_filter_fails_on_inplace_update_of_stage(self):
         with TestDir() as test_dir:
             dataset_url = osp.join(test_dir, "dataset")
             dataset = Dataset.from_iterable(
@@ -84,11 +70,11 @@ class TransformTest(TestCase):
             with self.subTest("without overwrite"):
                 run(
                     self,
-                    "transform",
+                    "filter",
                     "-p",
                     project_dir,
-                    "-t",
-                    "random_split",
+                    "-e",
+                    "/item",
                     "HEAD:source-1",
                     expected_code=1,
                 )
@@ -97,11 +83,11 @@ class TransformTest(TestCase):
                 with self.assertRaises(ReadonlyDatasetError):
                     run(
                         self,
-                        "transform",
+                        "filter",
                         "-p",
                         project_dir,
                         "--overwrite",
-                        "-t",
-                        "random_split",
+                        "-e",
+                        "/item",
                         "HEAD:source-1",
                     )
