+++ tests/integration/cli/test_generator.py
@@ -6,8 +6,7 @@ import pytest
 
 import datumaro.util.image as image_module
 
-from ..requirements import Requirements, mark_requirement
-
+from tests.requirements import Requirements, mark_requirement
 from tests.utils.test_utils import TestDir
 from tests.utils.test_utils import run_datum as run
