+++ tests/integration/cli/test_image_zip_format.py
@@ -8,8 +8,7 @@ import numpy as np
 from datumaro.components.dataset import Dataset, DatasetItem
 from datumaro.components.media import Image
 
-from ..requirements import Requirements, mark_requirement
-
+from tests.requirements import Requirements, mark_requirement
 from tests.utils.test_utils import TestDir, compare_datasets
 from tests.utils.test_utils import run_datum as run
