+++ tests/integration/cli/test_project.py
@@ -11,8 +11,7 @@ from datumaro.components.media import Image
 from datumaro.components.project import Project
 from datumaro.util.scope import scope_add, scoped
 
-from ..requirements import Requirements, mark_requirement
-
+from tests.requirements import Requirements, mark_requirement
 from tests.utils.test_utils import TestDir, compare_datasets, compare_dirs
 from tests.utils.test_utils import run_datum as run
