+++ tests/integration/cli/test_sly_point_cloud_format.py
@@ -6,8 +6,7 @@ from datumaro.components.dataset import Dataset
 from datumaro.components.extractor import DatasetItem
 from datumaro.components.media import Image, PointCloud
 
-from ..requirements import Requirements, mark_requirement
-
+from tests.requirements import Requirements, mark_requirement
 from tests.utils.test_utils import TestDir, compare_datasets_3d
 from tests.utils.test_utils import run_datum as run
